Provide audio format to pipe output command
So far, the pipe output of MPD had the problem that the commands
in the pipe had no information about the format of the raw audio data
sent into the pipe.

With this patch mpd substitutes the strings "$MPDPIPE_BITS",
"$MPDPIPE_CHANNELS", "$MPDPIPE_RATE", "$MPDPIPE_FORMAT"
within the command started for the pipe by the number of bits per sample,
the number of channels, the sample rate, and the sample format (one of
"S8", "S16", "S24_P32" or "S32"), respectively.

Example: if in mpd.conf an audio_output of type pipe contains the line

        command         "/usr/local/bin/mympdpipe $MPDPIPE_BITS $MPDPIPE_CHANNELS $MPDPIPE_RATE $MPDPIPE_FORMAT"

then mpd starts a command like
        /usr/local/bin/mympdpipe 16 2 44100 S16
for this pipe.