storage, db, mixer, command: remove more bogus "pure" attributes
This commit is similar to 788e3b31e1ab8243386339eaf136ede18f015d55,
and removes more "pure" attributes which were placed on functions that
could throw exceptions, which is illegal according to clang's
understanding of the attribute (but not according to GCC's).  GitHub
issue #58 was most likely about StorageDirectoryReader::GetInfo() and
Storage::GetInfo(), which still had "pure" attributes.

Closes #58