SocketUtil: explicitly set V6ONLY socket option
Currently, MPD emits the following warning on Linux when
`net.ipv6.bindv6only` is set to the default value zero:

	exception: bind to '0.0.0.0:6600' failed (continuing anyway,
	because binding to '[::]:6600' succeeded): Failed to bind
	socket: Address already in use

This is a known issues (#673, #557, …) caused by the fact that MPD, by
default, creates separate sockets for `AF_INET` and `AF_INET6`. With
this architecture it is not necessary to use the `AF_INET6` socket for
both IPv6 and IPv4. For this reason, explicitly enable the V6ONLY
socket option for `AF_INET6` sockets, thereby preventing the error
causing the warning to be emitted.