SongPrint: include `real_uri` in `song_print_info`
CUE virtual playlists is currently implemented by accessing ranges in
file referenced by `real_uri` locally. But the `real_uri` is not
included in the song info protocol, thus `ProxyDatabasePlugin` is not
able to retreive this information and access the corresponding file,
leads to not working cue tracks in satellite setup.

This commit adds the `real_uri` field into server-printed song info,
with the key name `RealUri`.