android: Add functions to convert DetachedSong to Java SongInfo object
This moves all the song tags into the hashmap and moves that into the song info object along with duration and the song uri.

It makes sure to clean up local references except for the values passed through in return.

In the future we could prefetch or cache the jni field IDs to get a bit of a performance improvement.