Make suffixes functions ignore case
Previously, when using a suffixes function in a decoder plugin, suffixes
were directly compared in a case-sensitive manner, so a file ending in
e.g. .mp3 would not be treated the same as .MP3, and then no decoder
plugin might be found for the .MP3 file.

This commit introduces a comparator that makes comparisons ignore case
by always lowering the case on both inputs.

The long and tiring std::set<std::string, std::less<>> is replaced with
DecoderPlugin::SuffixSet so suffixes functions no longer have to depend
on an arbitrary type that may change in the future, and can from now on
use this abstraction.

Credit for the comparator goes to:
Kevin "Alipha" Spinar, from #C++ on libera.chat