Logging: added configurable log timestamp options.
Log messages are prefixed with a timestamp when
written to a log file or to stdout (but not when
written to syslog or to systemd log).

The default timestamp is in seconds.
It can be configured by setting log_timestamp to
one of: none, minutes, seconds, or milliseconds.

Example:
none:         "message" (without timestamp)
minutes:      "Jan 30 17:11 : message"
seconds:      "Jan 30 17:11:32 : message"
milliseconds: "Jan 30 17:11:32.271 : message"

Having milliseconds resolution in the log gives more
information than just seconds when debugging or analyzing
response time issues or effects of changes in server or
client code.