output/Thread: keep track of the `AudioOutputSource` state
Add a `source_state` field which is an enum describing the state of
the `AudioOutputSource`.

This is important because the `AudioOutputSource` must not be used
after it has been flushed (because Filter::Flush() invalidates the
`Filter` object).  Therefore, with this new field, we avoid calling
the methods Fill(), Cancel() etc. which would otherwise crash (with
assertion failure in AudioOutputSource or nullptr dereference in
TwoFilters::Reset()).

Closes https://github.com/MusicPlayerDaemon/MPD/issues/2232