output/Source: add field `filter_flushed`
Skip the Filter::Reset() call after Flush() has already been called
and add a few assertions ensuring that Fill() is never called after
Flush().  For that, a new Filter would need to be created.

This allows checking in Open() whether reusing the existing filter is
possible.  If `filter_flushed` is true, we need to close the filter
and open a new one.

And this amplifies bug
https://github.com/MusicPlayerDaemon/MPD/issues/2232 by detecting the
condition that crashes more often.