output/oss: handle EAGAIN
Commit 9a7a8ec137fbe added an implicit `O_NONBLOCK` to all open()
calls done by MPD, which is a good idea (to avoid open() to block),
but it has side effects for OSS devices: a write() is now really
non-blocking.

We could make the OSS output change the file descriptor back to
blocking mode, which would be a simple one-liner.  But instead, I
decided to call poll() on EAGAIN, because it might allow other
interesting features, such as implementing Interrupt() for devices
that block for a long time (not part of this commit, but might come
soon).

Closes https://github.com/MusicPlayerDaemon/MPD/issues/2263