Add: test program to calculate the inserted audio CD ID.
Generates a CDID, following Musicbrainz instructions, only using cdio
for finding track offsets (slow, will read the disc a each call).

Requires a disc drive connected with an audio CD in it.
Calling "./test/MusicbrainzCDID cdda:///1" will output the ID found in stdout.

It is possible to chain it with the previous test program:
./test/MusicbrainzFetch `./test/MusicbrainzCDID cdda:///1`