output/Thread: reopen output after flush+pause ("single" mode)
When playback is paused at the end of a track due to "single" mode,
the source is first flushed and the output is drained (both does not
happen when pausing manually).  This sets source_state=FLUSHED which
means playback cannot be resumed until the AudioOutputSource is
reopened.  This however does not happen because
AudioOutputControl::Open() thinks it is not necessary; this method
however did not consider the AudioOutputSource flush state.  The
solution is to add yet another flag called "should_reopen" which is
set whenever the source is flushed.

Closes https://github.com/MusicPlayerDaemon/MPD/issues/2279