output/PipeWire: fix resuming playback after it was paused by "single" mode
This fixes another cause for the "single" mode drain/resume bug.  The
first (generic) cause was fixed by commit 9ab928a03668, and this is
the PipeWire-specific cause for another bug with the same symptoms:

When draining the PipeWire output at the end of a song in "single"
mode, PipeWireOutput::Process() returns without queueing the buffer
that was allocated by pw_stream_dequeue_buffer().  This leaves
libpipewire stuck and playback cannot be resumed.

The solution is to move the `drain_requested` check and the
pw_stream_flush() to the beginning of the method before the
pw_stream_dequeue_buffer() call.  That way, we can return without
still owning a buffer that we will never give back.

Closes https://github.com/MusicPlayerDaemon/MPD/issues/2315