lib/xiph/VorbisPicture: remove METADATA_BLOCK_PICTURE size limit
This limit was implemented to protect MPD against huge allocations,
but the limit was never actually useful, because at this point, it was
already too late: the original (base64-encoded) tag was already loaded
into memory.  If we want to limit tag sizes, we should instead
implement that prior to loading the Ogg packet into memory.

Closes https://github.com/MusicPlayerDaemon/MPD/issues/2352